+++ app/models/referred_by.rb
@@ -1,8 +1,11 @@
-class ReferredBy < ActiveRecord::Base
-  has_many :payload_requests
 
-  validates :root, presence: true
-  validates :path, presence: true
+module RushHour
+  class ReferredBy < ActiveRecord::Base
+    has_many :payload_requests
+
+    validates :root, presence: true
+    validates :path, presence: true
 
 
+  end
 end
