+++ app/models/resolution.rb
@@ -1,7 +1,10 @@
-class Resolution < ActiveRecord::Base
-  has_many :payload_requests
-  validates :height, presence: true
-  validates :width, presence: true
+module RushHour
+
+  class Resolution < ActiveRecord::Base
+    has_many :payload_requests
+    validates :height, presence: true
+    validates :width, presence: true
 
 
+  end
 end
