+++ app/models/user_agent.rb
@@ -1,7 +1,10 @@
-class UserAgent < ActiveRecord::Base
-  has_many :payload_requests
-  validates :os, presence: true
+module RushHour
 
-  validates :browser, presence: true
+  class UserAgent < ActiveRecord::Base
+    has_many :payload_requests
+    validates :os, presence: true
 
+    validates :browser, presence: true
+
+  end
 end
